caxselling/bor/src/requirements.txt
pydantic == 2.8.2
pyYAML == 6.0.2
flask-restx == 1.3.0
flask == 2.2.5
Flask-Cors == 5.0.0
swagger-ui-bundle == 0.0.5
python_dateutil == 2.8.2
paho-mqtt == 1.5.1
requests
setuptools >= 75.8.0 
jinja2
cython
pillow == 11.2.1